// backend/src/common/enums/matching.entity.enum.ts
/**
 * Available entities for fields matching.
 */
enum MatchingEntityEnum {
    CUSTOMER = 'CUSTOMER',
    CREDENTIALS = 'CREDENTIALS',
    LEAD = 'LEAD',
    CATEGORY = 'CRM_CATEGORY',
    OBJECT = 'OBJECT',
    CHILDREN = 'CHILDREN',
    MONTHLY_BUDGET = 'MONTHLY_BUDGET',
    JOB = 'JOB',
    PASSPORT = 'PASSPORT',
    APPLICANT = 'APPLICANT',
    BANK_ACCOUNT = 'BANK_ACCOUNT',
    EXISTING_PROPERTY = 'EXISTING_PROPERTY',
    APPLICATION = 'APPLICATION',
    DOCUMENT = 'DOCUMENT'
}
export default MatchingEntityEnum;